backend/requirements.txt
fastapi==0.115.6
pydantic==2.10.4
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-dotenv==1.0.1
google-genai>=1.0.0
PyMuPDF==1.25.3
python-docx==1.1.2
scikit-learn==1.6.1
scipy==1.15.1
pandas==2.2.3
numpy==2.2.2
reportlab==4.2.5
